Relocated get_xyz_string and get_xyz_matrix
From Species to Parser
to avoid circular import (plus, it makes sense)
