+++ arc/ts/run_autotst.py
@@ -19,7 +19,7 @@ import logging
 
 from autotst.reaction import AutoTST_Reaction
 
-from arc.species import get_xyz_string
+from arc.parser import get_xyz_string
 from arc.exceptions import TSError
 from arc.settings import arc_path
